refactor: extract RuntimeImpl/ExecutorImpl from rpc/mod.rs to cell_launcher
Phase 0.1 of the cached-cargo-workspace split. Breaks the cycle where
src/rpc/mod.rs imported cell::proc::{Builder,FuelEstimator} (rpc -> cell)
while cell::executor imported rpc::create_runtime_client (cell -> rpc).

Both directions of the cycle ran through wiring code that didn't belong in
either module. The capnp Server impls for system_capnp::{runtime,executor}
construct cells from RPC requests -- that's wiring, not protocol -- and now
live at the bin level in src/cell_launcher.rs.

Files moved out of rpc/mod.rs (-409 lines):
- RuntimeImpl + impl Server + make_executor + create_runtime_client
- ExecutorImpl + impl Server (the spawn handler with FuelEstimator)
- read_text_list/read_text_list_result/read_data_result helpers
- MAX_WASM_BYTES const

After this commit, src/rpc/* references zero items from src/cell/*.
src/cell/executor.rs still references rpc types -- the remaining half of
the cycle is Phase 0.2's job (lifting cell/executor.rs out of cell/).

No behavior change. cargo check --lib --bins --tests -p ww passes clean.
8 test/example callers migrated from ww::rpc::create_runtime_client to
ww::cell_launcher::create_runtime_client.
