+++ tests/snap_hello_rs_http_listener_e2e.rs
@@ -18,8 +18,9 @@
 
 use tokio::sync::{mpsc, oneshot, watch};
 
+use ww::cell_launcher::create_runtime_client;
 use ww::dispatcher::server::{new_registry, CgiRequest, CgiResponse};
-use ww::rpc::{create_runtime_client, CachePolicy, NetworkState};
+use ww::rpc::{CachePolicy, NetworkState};
 use ww::system_capnp;
 
 const SNAP_WASM_PATH: &str = "examples/snap-hello-rs/bin/snap-hello-rs.wasm";
