+++ tests/status_cell_e2e.rs
@@ -15,8 +15,9 @@
 
 use tokio::sync::{mpsc, watch};
 
+use ww::cell_launcher::create_runtime_client;
 use ww::dispatcher::wagi;
-use ww::rpc::{create_runtime_client, CachePolicy, NetworkState};
+use ww::rpc::{CachePolicy, NetworkState};
 
 const STATUS_WASM_PATH: &str = "std/status/bin/status.wasm";
