+++ tests/status_cell_http_listener_e2e.rs
@@ -23,8 +23,9 @@
 
 use tokio::sync::{mpsc, oneshot, watch};
 
+use ww::cell_launcher::create_runtime_client;
 use ww::dispatcher::server::{new_registry, CgiRequest};
-use ww::rpc::{create_runtime_client, CachePolicy, NetworkState};
+use ww::rpc::{CachePolicy, NetworkState};
 use ww::system_capnp;
 
 const STATUS_WASM_PATH: &str = "std/status/bin/status.wasm";
