+++ tests/stdin_shutdown_integration.rs
@@ -18,7 +18,8 @@ use capnp_rpc::RpcSystem;
 use tokio::sync::mpsc;
 use tokio_util::compat::TokioAsyncReadCompatExt;
 
-use ww::rpc::{create_runtime_client, CachePolicy, NetworkState};
+use ww::cell_launcher::create_runtime_client;
+use ww::rpc::{CachePolicy, NetworkState};
 use ww::system_capnp;
 
 const ECHO_WASM_PATH: &str = "examples/echo/bin/echo.wasm";
